+++ src/services/Drafts.php
@@ -19,6 +19,7 @@ use craft\events\DraftEvent;
 use craft\helpers\DateTimeHelper;
 use craft\helpers\Db;
 use craft\helpers\ElementHelper;
+use Throwable;
 use yii\base\Component;
 use yii\base\Exception;
 use yii\base\InvalidArgumentException;
@@ -37,24 +38,24 @@ class Drafts extends Component
     /**
      * @event DraftEvent The event that is triggered before a draft is created.
      */
-    const EVENT_BEFORE_CREATE_DRAFT = 'beforeCreateDraft';
+    public const EVENT_BEFORE_CREATE_DRAFT = 'beforeCreateDraft';
 
     /**
      * @event DraftEvent The event that is triggered after a draft is created.
      */
-    const EVENT_AFTER_CREATE_DRAFT = 'afterCreateDraft';
+    public const EVENT_AFTER_CREATE_DRAFT = 'afterCreateDraft';
 
     /**
      * @event DraftEvent The event that is triggered before a draft is published.
      * @since 3.6.0
      */
-    const EVENT_BEFORE_APPLY_DRAFT = 'beforeApplyDraft';
+    public const EVENT_BEFORE_APPLY_DRAFT = 'beforeApplyDraft';
 
     /**
      * @event DraftEvent The event that is triggered after a draft is applied to its canonical element.
      * @see applyDraft()
      */
-    const EVENT_AFTER_APPLY_DRAFT = 'afterApplyDraft';
+    public const EVENT_AFTER_APPLY_DRAFT = 'afterApplyDraft';
 
     /**
      * @var Connection|array|string The database connection to use
@@ -108,17 +109,16 @@ class Drafts extends Component
      * @param array $newAttributes any attributes to apply to the draft
      * @param bool $provisional Whether to create a provisional draft
      * @return ElementInterface The new draft
-     * @throws \Throwable
+     * @throws Throwable
      */
     public function createDraft(
         ElementInterface $canonical,
-        int              $creatorId,
-        ?string          $name = null,
-        ?string          $notes = null,
-        array            $newAttributes = [],
-        bool             $provisional = false
-    ): ElementInterface
-    {
+        int $creatorId,
+        ?string $name = null,
+        ?string $notes = null,
+        array $newAttributes = [],
+        bool $provisional = false
+    ): ElementInterface {
         // Make sure the canonical element isn't a draft or revision
         if ($canonical->getIsDraft() || $canonical->getIsRevision()) {
             throw new InvalidArgumentException('Cannot create a draft from another draft or revision.');
@@ -160,7 +160,7 @@ class Drafts extends Component
             $draft = Craft::$app->getElements()->duplicateElement($canonical, $newAttributes);
 
             $transaction->commit();
-        } catch (\Throwable $e) {
+        } catch (Throwable $e) {
             $transaction->rollBack();
             throw $e;
         }
@@ -215,7 +215,7 @@ class Drafts extends Component
      * @param string|null $notes
      * @param bool $markAsSaved
      * @return bool
-     * @throws \Throwable
+     * @throws Throwable
      */
     public function saveElementAsDraft(ElementInterface $element, ?int $creatorId = null, ?string $name = null, ?string $notes = null, bool $markAsSaved = true): bool
     {
@@ -245,7 +245,7 @@ class Drafts extends Component
      *
      * @param ElementInterface $draft The draft
      * @return ElementInterface The canonical element with the draft applied to it
-     * @throws \Throwable
+     * @throws Throwable
      * @since 3.6.0
      */
     public function applyDraft(ElementInterface $draft): ElementInterface
@@ -314,13 +314,24 @@ class Drafts extends Component
 
                 // We still need to validate so the SlugValidator gets run
                 $draft->setScenario(Element::SCENARIO_ESSENTIALS);
+                $draft->validate();
+
+                // If there are any errors on the URI, re-validate as disabled
+                if ($draft->hasErrors('uri') && $draft->enabled) {
+                    $draft->enabled = false;
+                    $draft->validate();
+                }
+
+                if ($draft->hasErrors()) {
+                    throw new InvalidElementException($draft, 'Draft ' . $draft->id . ' could not be applied because it doesn\'t validate.');
+                }
 
                 try {
-                    $elementsService->saveElement($draft);
+                    $elementsService->saveElement($draft, false);
                     Db::delete(Table::DRAFTS, [
                         'id' => $draftId,
                     ]);
-                } catch (\Throwable $e) {
+                } catch (Throwable $e) {
                     // Put everything back
                     $draft->draftId = $draftId;
                     $draft->attachBehavior('draft', $behavior);
@@ -333,7 +344,7 @@ class Drafts extends Component
             }
 
             $transaction->commit();
-        } catch (\Throwable $e) {
+        } catch (Throwable $e) {
             $transaction->rollBack();
 
             if ($e instanceof InvalidElementException) {
@@ -390,7 +401,7 @@ class Drafts extends Component
             $draft = $elementType::find()
                 ->draftId($draftInfo['draftId'])
                 ->status(null)
-                ->siteId('*')
+                ->site('*')
                 ->one();
 
             if ($draft) {
@@ -424,12 +435,11 @@ class Drafts extends Component
     public function insertDraftRow(
         ?string $name,
         ?string $notes = null,
-        ?int    $creatorId = null,
-        ?int    $canonicalId = null,
-        bool    $trackChanges = false,
-        bool    $provisional = false
-    ): int
-    {
+        ?int $creatorId = null,
+        ?int $canonicalId = null,
+        bool $trackChanges = false,
+        bool $provisional = false
+    ): int {
         Db::insert(Table::DRAFTS, [
             'canonicalId' => $canonicalId,
             'creatorId' => $creatorId,
